+++ packages/t-chart/src/index.js
@@ -1,0 +1,6 @@
+export * from "./coordinate";
+export * from "./geometry";
+export * from "./guide";
+export * from "./plot";
+export * from "./scale";
+export * from "./statistic";