+++ packages/t-chart/vite.config.js
@@ -1,0 +1,30 @@
+import { resolve } from "node:path";
+import { fileURLToPath, URL } from "node:url";
+import { defineConfig } from "vite";
+
+export default defineConfig({
+  resolve: {
+    alias: {
+      "@": fileURLToPath(new URL("./src", import.meta.url)),
+    },
+  },
+  build: {
+    lib: {
+      entry: resolve(__dirname, "src/index.js"),
+      name: "@tinyvis/t-chart",
+      fileName: (format) => {
+        if (format === "es")
+          return "index.js";
+        if (format === "cjs")
+          return "index.cjs";
+      },
+      formats: ["es", "cjs"],
+    },
+    rollupOptions: {
+      external: [],
+      output: {
+        globals: {},
+      },
+    },
+  },
+});