+++ packages/t-svg/eslint.config.mjs
@@ -1,3 +1,3 @@
-import { config } from "@TRender/eslint-config/svg";
+import { config } from "@tinyvis/eslint-config/svg";
 
 export default config;