+++ apps/playground/CHANGELOG.md
@@ -1,8 +1,8 @@
-# @TRender/playground
+# @tinyvis/playground
 
 ## 0.0.1
 
 ### Patch Changes
 
-- Updated dependencies [[`2616b47`](https://github.com/taiiiyang/TRender/commit/2616b47cb930c199bbe221ba42759cd78395e703), [`dea474f`](https://github.com/taiiiyang/TRender/commit/dea474f956f6dcddf2481a7ed492b5c2ff0144b2), [`3813518`](https://github.com/taiiiyang/TRender/commit/3813518a82e0fc5af6e715c4ce8aa676b5fa81d1)]:
-  - @TRender/t-svg@0.0.1
+- Updated dependencies [[`2616b47`](https://github.com/taiiiyang/TinyVis/commit/2616b47cb930c199bbe221ba42759cd78395e703), [`dea474f`](https://github.com/taiiiyang/TinyVis/commit/dea474f956f6dcddf2481a7ed492b5c2ff0144b2), [`3813518`](https://github.com/taiiiyang/TinyVis/commit/3813518a82e0fc5af6e715c4ce8aa676b5fa81d1)]:
+  - @tinyvis/t-svg@0.0.1